requirements.txt
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
transformers>=4.20.0
PyYAML>=6.0
tqdm>=4.64.0
scipy>=1.7.0
matplotlib>=3.4.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
pandas>=1.3.0
h5py>=3.6.0
protobuf<=3.20.3
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
pyopengl>=3.1.0